add docs for GASTBuilder
