+++ Sources/Common/Docs/FileProvider.md
@@ -1,6 +1,6 @@
 # FileProvider
 
-Interface for object which can deal with files in file system
+Interface for object which is can deal with files in file system
 
 ``` swift
 public protocol FileProvider
