+++ Sources/Common/Docs/_Footer.md
@@ -1,1 +1,1 @@
-Generated at 2021-01-05T09:53:46+0300 using [swift-doc](https://github.com/SwiftDocOrg/swift-doc) 1.0.0-beta.5.
+Generated at 2021-01-05T11:18:56+0300 using [swift-doc](https://github.com/SwiftDocOrg/swift-doc) 1.0.0-beta.5.
