+++ Sources/GASTBuilder/Builders/AnyMediaTypesBuilder.swift
@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  AnyMediaTypesBuilder.swift
 //  
 //
 //  Created by Александр Кравченков on 27.12.2020.
@@ -10,16 +10,25 @@ import Swagger
 import GASTTree
 import Common
 
+/// Just an interface for any GAST-MediaType builder
 public protocol MediaTypesBuilder {
     func buildMediaItems(items: [String: MediaItem]) throws -> [MediaTypeObjectNode]
 }
 
+/// Default implementation of MediaTypesBuilder
+///
+/// ## Don't support
+///
+/// ### MediaType content must be a `reference`. Other in-place declaration itsn't supported. Because it that case we can't generate model
 public struct AnyMediaTypesBuilder: MediaTypesBuilder {
 
     public let schemaBuilder: SchemaBuilder
     /// If set to `false` disable errors for cases when MediaType schema cotains definition of object/enum/alias e.t.c
     /// If set to `true` throws error for any case except reference
     /// By default set to `true`
+    ///
+    /// **WARNING**
+    /// Don't touch it in SurfGen code
     public let enableDisclarationChecking: Bool
 
     public init(schemaBuilder: SchemaBuilder, enableDisclarationChecking: Bool = true) {
