+++ Sources/GASTBuilder/Builders/AnyParametersBuilder.swift
@@ -10,10 +10,39 @@ import Common
 import Swagger
 import GASTTree
 
+/// Just an interface for any GAST-Parameter builder
 public protocol ParametersBuilder {
+    /// Can build `parameter`
     func build(parameters: [ComponentObject<Parameter>]) throws -> [ParameterNode]
 }
 
+/// Default implementation for `ParametersBuilder`.
+/// Can build parameters both from `components.parameters` and from `paths.operations.parameters`
+/// 
+///
+/// ## Don't support
+///
+/// ### Content in paramter's type. You cant declare `schema` in parameter's type
+///
+/// For example it's **ok**:
+///
+/// ```YAML
+///  Param1:
+///     type: integer
+///
+///  Param2:
+///     type:
+///         $ref: "..."
+/// ```
+/// But it's **not**
+///
+/// ```YAML
+///  Param1:
+///     type:
+///         content:
+///             type: object
+///             ....
+/// ```
 public struct AnyParametersBuilder: ParametersBuilder {
 
     private let schemaBuilder: SchemaBuilder
