+++ Sources/GASTBuilder/Builders/AnyResponsesBuilder.swift
@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  AnyResponsesBuilder.swift
 //  
 //
 //  Created by Александр Кравченков on 27.12.2020.
@@ -10,14 +10,21 @@ import GASTTree
 import Swagger
 import Common
 
+/// Just an interface for any GAST-response builder
 public protocol ResponseBuilder {
     func build(response: Response) throws -> ResponseNode
 }
 
+/// The same as `ResponseBuilder` but can build `components.responses`
 public protocol ResponsesBuilder {
     func build(responses: [ComponentObject<Response>]) throws -> [ComponentResponseNode]
 }
 
+/// Default implementation of both `ResponseBuilder` and `ResponsesBuilder`
+///
+/// ## Don't support
+///
+/// ### Response body without content. They are mapped to `nil`
 public struct AnyResponsesBuilder: ResponsesBuilder, ResponseBuilder {
 
     public let mediaTypesBuilder: MediaTypesBuilder
