+++ Sources/GASTBuilder/Builders/AnyServiceBuilder.swift
@@ -10,10 +10,23 @@ import Common
 import Swagger
 import GASTTree
 
+/// Just an interface for any GAST-Service builder
 public protocol ServiceBuilder {
+    /// Build all item which are under `paths:`
     func build(paths: [Path]) throws -> [PathNode]
 }
 
+/// Default implementation for `ServiceBuilder`
+/// Builds `path` elements of Open-API spec
+///
+/// **WARNING**
+///
+/// This struct cuts out response bodies without content. So if response doesn't have body then it will be just `nil`
+///
+/// - See: https://swagger.io/docs/specification/paths-and-operations/
+///
+/// - Todo: Parameters declared `in-pace` have `name` property set to empty string. May be it isn't good idea. It isn't affect code generation.
+/// - Todo: In Swagger exists response key `default`. It's response wihout specific `httpCode`. Now we support it. But to honest. I don't like it. May be we need to throw a warning about this.
 public struct AnyServiceBuilder: ServiceBuilder {
 
     let parameterBuilder: ParametersBuilder
@@ -31,6 +44,7 @@ public struct AnyServiceBuilder: ServiceBuilder {
         self.responseBuilder = responseBuilder
     }
 
+    /// Build all item which are under `paths:`
     public func build(paths: [Path]) throws -> [PathNode] {
         return try paths.map { path in
             let operations = try wrap(self.build(operations: path.operations),
