+++ Sources/GASTBuilder/Docs/AnyResponsesBuilder.md
@@ -1,0 +1,45 @@
+# AnyResponsesBuilder
+
+Default implementation of both `ResponseBuilder` and `ResponsesBuilder`
+
+``` swift
+public struct AnyResponsesBuilder: ResponsesBuilder, ResponseBuilder
+```
+
+## Don't support
+
+### Response body without content. They are mapped to `nil`
+
+## Inheritance
+
+[`ResponseBuilder`](./ResponseBuilder), [`ResponsesBuilder`](./ResponsesBuilder)
+
+## Initializers
+
+### `init(mediaTypesBuilder:)`
+
+``` swift
+public init(mediaTypesBuilder: MediaTypesBuilder)
+```
+
+## Properties
+
+### `mediaTypesBuilder`
+
+``` swift
+let mediaTypesBuilder: MediaTypesBuilder
+```
+
+## Methods
+
+### `build(responses:)`
+
+``` swift
+public func build(responses: [ComponentObject<Response>]) throws -> [ComponentResponseNode]
+```
+
+### `build(response:)`
+
+``` swift
+public func build(response: Response) throws -> ResponseNode
+```
