+++ Sources/GASTBuilder/Docs/GASTBuilder.md
@@ -1,0 +1,15 @@
+# GASTBuilder
+
+Just an interface for any `GAST` builder
+
+``` swift
+public protocol GASTBuilder
+```
+
+## Requirements
+
+### build(filePath:​)
+
+``` swift
+func build(filePath: String) throws -> RootNode
+```
