+++ Sources/GASTBuilder/Docs/MediaTypesBuilder.md
@@ -1,0 +1,15 @@
+# MediaTypesBuilder
+
+Just an interface for any GAST-MediaType builder
+
+``` swift
+public protocol MediaTypesBuilder
+```
+
+## Requirements
+
+### buildMediaItems(items:​)
+
+``` swift
+func buildMediaItems(items: [String: MediaItem]) throws -> [MediaTypeObjectNode]
+```
