+++ Sources/GASTBuilder/Docs/RequestBodyBuilder.md
@@ -1,0 +1,15 @@
+# RequestBodyBuilder
+
+Just an interface for any GAST-RequestBody builder
+
+``` swift
+public protocol RequestBodyBuilder
+```
+
+## Requirements
+
+### build(requestBody:​)
+
+``` swift
+func build(requestBody: RequestBody) throws -> RequestBodyNode
+```
