+++ Sources/GASTBuilder/Docs/ResponseBuilder.md
@@ -1,0 +1,15 @@
+# ResponseBuilder
+
+Just an interface for any GAST-response builder
+
+``` swift
+public protocol ResponseBuilder
+```
+
+## Requirements
+
+### build(response:​)
+
+``` swift
+func build(response: Response) throws -> ResponseNode
+```
