+++ Sources/GASTBuilder/Docs/ResponsesBuilder.md
@@ -1,0 +1,15 @@
+# ResponsesBuilder
+
+The same as `ResponseBuilder` but can build `components.responses`
+
+``` swift
+public protocol ResponsesBuilder
+```
+
+## Requirements
+
+### build(responses:​)
+
+``` swift
+func build(responses: [ComponentObject<Response>]) throws -> [ComponentResponseNode]
+```
