+++ Sources/GASTBuilder/Docs/_Sidebar.md
@@ -1,0 +1,27 @@
+<details>
+<summary>Types</summary>
+
+  - [AnyGASTBuilder](./AnyGASTBuilder)
+  - [AnyMediaTypesBuilder](./AnyMediaTypesBuilder)
+  - [AnyParametersBuilder](./AnyParametersBuilder)
+  - [AnyRequestBodiesBuilder](./AnyRequestBodiesBuilder)
+  - [AnyResponsesBuilder](./AnyResponsesBuilder)
+  - [AnySchemaBuilder](./AnySchemaBuilder)
+  - [AnyServiceBuilder](./AnyServiceBuilder)
+
+</details>
+
+<details>
+<summary>Protocols</summary>
+
+  - [GASTBuilder](./GASTBuilder)
+  - [MediaTypesBuilder](./MediaTypesBuilder)
+  - [ParametersBuilder](./ParametersBuilder)
+  - [RequestBodiesBuilder](./RequestBodiesBuilder)
+  - [RequestBodyBuilder](./RequestBodyBuilder)
+  - [ResponseBuilder](./ResponseBuilder)
+  - [ResponsesBuilder](./ResponsesBuilder)
+  - [SchemaBuilder](./SchemaBuilder)
+  - [ServiceBuilder](./ServiceBuilder)
+
+</details>
