+++ Sources/GASTBuilder/Extensions/Swagger.swift
@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  Swagger.swift
 //  
 //
 //  Created by Александр Кравченков on 14.12.2020.
@@ -10,41 +10,12 @@ import Swagger
 import Common
 import GASTTree
 
-extension Schema {
-    func extractType() throws -> PropertyNode.PossibleType {
-        switch self.type {
-        case .any:
-            throw CommonError(message: "Type is `any`, but we can process only primitive types, arrays and $ref")
-        case .object:
-            throw CommonError(message: "Type is `object`, but we can process only primitive types, arrays and $ref")
-        case .group:
-            throw CommonError(message: "Type is `group`, but we can process only primitive types, arrays and $ref")
-        case .array(let arr):
-            switch arr.items {
-            case .multiple:
-                throw CommonError(message: "Array conains multiple items declaration. So we can't process it now")
-            case .single(let schema):
-                let type = try schema.extractType()
-                guard case .simple(let val) = type else {
-                    throw CommonError(message: "Array conains single item with wrong type \(type). But we can process only primitive types and $ref in this case")
-                }
-                return .array(.init(itemsType: val))
-            }
-        case .reference(let ref):
-            return .simple(.ref(ref.rawValue))
-        case .boolean:
-            return .simple(.entity(.boolean))
-        case .string:
-            return .simple(.entity(.string))
-        case .number:
-            return .simple(.entity(.number))
-        case .integer:
-            return .simple(.entity(.integer))
-        }
-    }
-}
-
 extension ParameterLocation {
+    /// Converts `Swagger` parameters location to `SurfGen`
+    ///
+    /// ## Don't support
+    ///
+    /// ### `cookie` and `header` as parameter's location
     func convert() throws -> ParameterNode.Location {
         switch self {
         case .query:
