+++ generate_documentation_for_specific_modules.sh
@@ -2,7 +2,7 @@ rm -rf Sources/Common/Docs
 rm -rf Sources/CodeGenerator/Docs
 # rm -rf Sources/ReferenceExtractor/Docs
 # rm -rf Sources/Pipelines/Docs
-# rm -rf Sources/GASTBuilder/Docs
+rm -rf Sources/GASTBuilder/Docs
 # rm -rf Sources/GASTTree/Docs
 
 rm Sources/CodeGenerator/README.md
@@ -12,11 +12,13 @@ swift doc generate Sources/Common --module-name SurfGen --output Sources/Common/
 swift doc generate Sources/CodeGenerator --module-name SurfGen --output Sources/CodeGenerator/Docs --base-url=./Docs
 # swift doc generate Sources/ReferenceExtractor --module-name SurfGen --output Sources/ReferenceExtractor/Docs --base-url=./
 # swift doc generate Sources/Pipelines --module-name SurfGen --output Sources/Pipelines/Docs --base-url=./
-# swift doc generate Sources/GASTBuilder --module-name SurfGen --output Sources/GASTBuilder/Docs --base-url=./
+swift doc generate Sources/GASTBuilder --module-name SurfGen --output Sources/GASTBuilder/Docs --base-url=./
 # swift doc generate Sources/GASTTree --module-name SurfGen --output Sources/GASTTree/Docs --base-url=./
 
 mv Sources/CodeGenerator/Docs/Home.md Sources/CodeGenerator/README.md 
 mv Sources/Common/Docs/Home.md Sources/Common/README.md 
+mv Sources/GASTBuilder/Docs/Home.md Sources/GASTBuilder/README.md 
 
 sed -i '' 's|):|.md):|g' Sources/CodeGenerator/README.md
 sed -i '' 's|):|.md):|g' Sources/Common/README.md
+sed -i '' 's|):|.md):|g' Sources/GASTBuilder/README.md
